+++ backend/src/main/java/com/sdacademy/budgettracker/repository/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Double> {
     Optional<User> findByUsernameAndPassword(String username, String password);
 }
